requirements/common.txt
Django==4.1.4
djangorestframework==3.14.0
argon2-cffi==21.2.0
python-dotenv==0.21.0
redis==4.4.0
PyJWT==2.6.0
kavenegar==1.1.2
Pillow==9.3.0
unicode-slugify==0.1.5
requests==2.28.1
